Backend/order_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
alembic
pydantic[email]
